Sprint 5
Added existing patient look up feature
Added many to one relationship between patient and visit
Added validation to new patient and find patient
Changed String dates to Date objects
Fixed css styling on patient page

+++ VaccinationTracker/src/main/java/vt/data/HealthRecordRepository.java
@@ -6,6 +6,6 @@ import vt.HealthRecord;
 
 //Author: Nick Riley
 
-public interface HealthRecordRepository extends JpaRepository<HealthRecord, Integer> {
+public interface HealthRecordRepository extends JpaRepository<HealthRecord, Long> {
 
 }
